+++ src/main/java/org/terasology/sample/nui/EnvironmentInfoScreen.java
