+++ src/main/java/org/terasology/sample/nui/FortuneTeller.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.nui;
+package org.terasology.sample.nui;
 
 
 import org.terasology.rendering.nui.CoreScreenLayer;
