+++ src/main/java/org/terasology/sample/nui/KonamiScreen.java
