+++ src/main/java/org/terasology/sample/nui/LatinQuotes.java
