+++ src/main/java/org/terasology/sample/nui/NeverGiveUp.java
