+++ src/main/java/org/terasology/sample/nui/TimeStatus.java
