+++ src/main/java/org/terasology/sample/nui/internal/TestListenerScreen.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.nui.internal;
+package org.terasology.sample.nui.internal;
 
 import org.terasology.nui.ListenerScreen;
