+++ src/main/java/org/terasology/sample/test/PrivateComponent.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014 MovingBlocks
+ * Copyright 2016 MovingBlocks
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,13 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.crash;
+package org.terasology.sample.test;
 
 import org.terasology.entitySystem.Component;
 
-/**
- * Used by a block meant to crash the game, yay!
- */
-public class CrashGameBlockComponent implements Component {
-    boolean reallyCrashy;
+public class PrivateComponent implements Component {
+    private PrivateComponent() {
+
+    }
 }
